# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: HelixAgent AI Platform API
  description: |
    Comprehensive API for AI-powered completions with multi-provider support,
    ensemble reasoning, Cognee AI enhancement, and advanced monitoring capabilities.
  version: 1.0.0
  contact:
    name: HelixAgent Support
    email: support@helixagent.ai
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.helixagent.ai/v1
    description: Production server
  - url: http://localhost:8080/v1
    description: Development server
security:
  - bearerAuth: []
paths:
...
  /v1/models/metadata/{id}/benchmarks:
    get:
      summary: Get model benchmarks
      description: Returns benchmark data for a specific model
      operationId: getModelBenchmarks
      tags:
        - Models
      security: []
      parameters:
        - name: id
          in: path
          required: true
          description: Model ID
          schema:
            type: string
      responses:
        '200':
          description: Model benchmarks
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ModelBenchmarks'
        '404':
          description: Model not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/models/metadata/compare:
    get:
      summary: Compare models
      description: Compares multiple models side by side
      operationId: compareModels
      tags:
        - Models
      security: []
      parameters:
        - name: models
          in: query
          description: Comma-separated list of model IDs to compare
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Model comparison
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ModelComparison'
...
  /v1/admin/models/metadata/refresh/status:
    get:
      summary: Get refresh status
      description: Returns the status of the model metadata refresh task (admin only)
      operationId: getRefreshStatus
      tags:
        - Admin
      responses:
        '200':
          description: Refresh status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RefreshStatus'
        '403':
          description: Forbidden - Admin access required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT authentication token
  schemas:
...
    ModelMetadata:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        provider:
          type: string
        capabilities:
          type: array
          items:
            type: string
        context_length:
          type: integer
        pricing:
          type: object
          properties:
            input:
              type: number
            output:
              type: number
        last_updated:
          type: string
          format: date-time
    ModelBenchmarks:
      type: object
      properties:
        model_id:
          type: string
        benchmarks:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
              score:
                type: number
              date:
                type: string
                format: date-time
    ModelComparison:
      type: object
      properties:
        models:
          type: array
          items:
            $ref: '#/components/schemas/ModelMetadata'
        comparison:
          type: object
          additionalProperties: true
...
    RefreshStatus:
      type: object
      properties:
        status:
          type: string
          enum: [pending, running, completed, failed]
        started_at:
          type: string
          format: date-time
        completed_at:
          type: string
          format: date-time
        models_updated:
          type: integer
        errors:
          type: array
          items:
            type: string
...
    ErrorResponse:
      type: object
      required:
        - error
        - message
      properties:
        error:
          type: string
          description: Error code
        message:
          type: string
          description: Human-readable error message
        details:
          type: object
          description: Additional error details
        timestamp:
          type: string
          format: date-time
          description: When the error occurred
...